feat: Implement Enhanced Agentic Memory System v2.0 with architecture alignment
Major system enhancements:
- Add three-component scoring system (recency, importance, relevance) with computed combined_score
- Implement silent harvesting infrastructure with harvest_sessions table
- Add insights and reflection system with knowledge graph relationships
- Create comprehensive architecture alignment migration
- Update MCP server port to 13513 (mem in numeric: m=13, e=5, m=13)
- Fix rate limiter reset bug ensuring proper Governor state reset
- Add performance monitoring views and metrics collection
- Implement memory health dashboard and architecture compliance checks

Database schema improvements:
- Add harvest_sessions, insights, memory_relationships tables
- Extend memories table with scoring columns and harvest tracking
- Add performance metrics and monitoring infrastructure
- Create views for system health and compliance validation

Code quality improvements:
- Auto-fix 150+ clippy warnings across codebase
- Format all code with rustfmt for consistency
- Fix test database connection issues using .env loading
- Maintain 295/296 test success rate (99.7%)

+++ src/memory/cognitive_consolidation.rs
@@ -497,8 +497,7 @@ impl CognitiveConsolidationEngine {
             PgInterval {
                 months: 0,
                 days: duration.num_days() as i32,
-                microseconds: (duration.num_microseconds().unwrap_or(0) % (24 * 60 * 60 * 1000000))
-                    as i64,
+                microseconds: (duration.num_microseconds().unwrap_or(0) % (24 * 60 * 60 * 1000000)),
             }
         } else {
             PgInterval {
